// app/backend/src/interfaces/LeaderbordInterface.ts
interface ILeaderBoard {
  name: string
  totalPoints: number
  totalGames: number
  totalVictories: number
  totalDraws: number
  totalLosses: number
  goalsFavor: number
  goalsOwn: number
  goalsBalance: number
  efficiency: number
}

interface IMatchLeaderBoard {
  id?: number;
  homeTeam: number;
  homeTeamGoals: number;
  awayTeam: number;
  awayTeamGoals: number;
  inProgress: boolean;
}

export {
  ILeaderBoard,
  IMatchLeaderBoard,
};